+++ src/main/resources/db/migration/V1_10__Create_transportations_table.sql
@@ -1,9 +1,13 @@
+# DROP SEQUENCE IF EXISTS `transportations_seq`;
+# CREATE SEQUENCE `transportations_seq`;
+
 DROP TABLE IF EXISTS `transportations`;
 
 CREATE TABLE `transportations` (
-   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-   `barco_id` int(10) NOT NULL,
-   `storehouse_id` int(10) NOT NULL,
+   #`id` BIGINT(21) DEFAULT nextval(`transportations_seq`),
+   `id` bigint(21) unsigned NOT NULL AUTO_INCREMENT,
+   `barco_id` bigint(10) NOT NULL,
+   `storehouse_id` bigint(10) NOT NULL,
    `distance` int(10) NOT NULL,
    `weight` int(10) NOT NULL,
    `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
