+++ src/main/resources/db/migration/V1_1__Create_cities_table.sql
@@ -1,7 +1,11 @@
+# DROP SEQUENCE IF EXISTS `cities_seq`;
+# CREATE SEQUENCE `cities_seq`;
+
 DROP TABLE IF EXISTS cities;
 
 CREATE TABLE `cities` (
-   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+   #`id` BIGINT(21) DEFAULT nextval(`cities_seq`),
+   `id` bigint(21) unsigned NOT NULL AUTO_INCREMENT,
    `title` varchar(255) CHARACTER SET utf8 COLLATE utf8_general_ci NULL,
    `description` varchar(500) CHARACTER SET utf8 COLLATE utf8_general_ci NULL,
    `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
