+++ src/main/resources/db/migration/V1_4__Create_barco_products_table.sql
@@ -1,9 +1,13 @@
+# DROP SEQUENCE IF EXISTS `barco_products_seq`;
+# CREATE SEQUENCE `barco_products_seq`;
+
 DROP TABLE IF EXISTS `barco_products`;
 
 CREATE TABLE `barco_products` (
-   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-   `barco_id` int(10) NOT NULL,
-   `product_id` int(10) NOT NULL,
+   #`id` BIGINT(21) DEFAULT nextval(`barco_products_seq`),
+   `id` bigint(21) unsigned NOT NULL AUTO_INCREMENT,
+   `barco_id` bigint(10) NOT NULL,
+   `product_id` bigint(10) NOT NULL,
    `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
    `updated_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
  PRIMARY KEY (`id`),
