+++ src/main/resources/db/migration/V1_6__Create_users_table.sql
@@ -1,13 +1,17 @@
+# DROP SEQUENCE IF EXISTS `users_seq`;
+# CREATE SEQUENCE `users_seq`;
+
 DROP TABLE IF EXISTS `users`;
 
 CREATE TABLE `users` (
-   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+   #`id` BIGINT(21) DEFAULT nextval(`users_seq`),
+   `id` bigint(21) unsigned NOT NULL AUTO_INCREMENT,
    `firstname` varchar(255) CHARACTER SET utf8 COLLATE utf8_general_ci NULL,
    `lastname` varchar(255) CHARACTER SET utf8 COLLATE utf8_general_ci NULL,
    `email` varchar(200) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
    `password` varchar(255) CHARACTER SET utf8 COLLATE utf8_general_ci NOT NULL,
    `bio` varchar(500) CHARACTER SET utf8 COLLATE utf8_general_ci NULL,
-   `position_id` int(10) NOT NULL,
+   `position_id` bigint(10) NOT NULL,
    `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
    `updated_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
  PRIMARY KEY (`id`),
