+++ src/main/resources/db/migration/V1_9__Create_storehouse_barcos_table.sql
@@ -1,9 +1,13 @@
+# DROP SEQUENCE IF EXISTS `storehouse_barcos_seq`;
+# CREATE SEQUENCE `storehouse_barcos_seq`;
+
 DROP TABLE IF EXISTS `storehouse_barcos`;
 
 CREATE TABLE `storehouse_barcos` (
-   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
-   `storehouse_id` int(10) NOT NULL,
-   `barco_id` int(10) NOT NULL,
+   #`id` BIGINT(21) DEFAULT nextval(`storehouse_barcos_seq`),
+   `id` bigint(21) unsigned NOT NULL AUTO_INCREMENT,
+   `storehouse_id` bigint(10) NOT NULL,
+   `barco_id` bigint(10) NOT NULL,
    `created_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
    `updated_at` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
  PRIMARY KEY (`id`),
